courses and teachers get fix, test for users, courses and teachers and division of tests into multiple files
