+++ src/api/endpoints/courses.py
@@ -7,16 +7,20 @@ from ..serializers import *
 
 @api_view(['GET'])
 def get_courses(request):
-    data = Course.objects.all()
-    print("Data:", data)
+    try:
+        data = Course.objects.all()
+    except Course.DoesNotExist:
+        return Response({"mensaje": "There are not courses yet"}, status=404)
     serializer = CourseSerializer(data, context={'request': request}, many=True)
-    print("Serializado: ", serializer)
     return Response(serializer.data)
 
 
 @api_view(['GET'])
 def get_course(request, pk=None):
-    data = Course.objects.get(pk=pk)
+    try:
+        data = Course.objects.get(pk=pk)
+    except Course.DoesNotExist:
+        return Response({"mensaje": "Course does not exist"}, status=404)
     serializer = CourseSerializer(data, context={'request': request})
     return Response(serializer.data)
